# frontend/data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BakeMate",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/reports/profit-and-loss": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Profit and Loss Report",
        "description": "Generate a Profit and Loss report for the specified period.\nOutput can be JSON, CSV, or PDF (PDF is placeholder).",
        "operationId": "get_profit_and_loss_report_api_v1_reports_profit_and_loss_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Start date for the report period (YYYY-MM-DD)",
              "title": "Start Date"
            },
            "description": "Start date for the report period (YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "End date for the report period (YYYY-MM-DD)",
              "title": "End Date"
            },
            "description": "End date for the report period (YYYY-MM-DD)"
          },
          {
            "name": "output_format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Output format for the report",
              "enum": [
                "json",
                "csv",
                "pdf"
              ],
              "default": "json",
              "title": "Output Format"
            },
            "description": "Output format for the report"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Profit And Loss Report Api V1 Reports Profit And Loss Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/sales-by-product": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Sales by Product Report",
        "description": "Generate a Sales by Product report for the specified period.\nOutput can be JSON, CSV, or PDF (PDF is placeholder).",
        "operationId": "get_sales_by_product_report_api_v1_reports_sales_by_product_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Start date for the report period (YYYY-MM-DD)",
              "title": "Start Date"
            },
            "description": "Start date for the report period (YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "End date for the report period (YYYY-MM-DD)",
              "title": "End Date"
            },
            "description": "End date for the report period (YYYY-MM-DD)"
          },
          {
            "name": "output_format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Output format for the report",
              "enum": [
                "json",
                "csv",
                "pdf"
              ],
              "default": "json",
              "title": "Output Format"
            },
            "description": "Output format for the report"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Sales By Product Report Api V1 Reports Sales By Product Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/ingredient-usage": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Ingredient Usage Report",
        "description": "Generate an Ingredient Usage report based on completed orders for the specified period.\nOutput can be JSON, CSV, or PDF (PDF is placeholder).",
        "operationId": "get_ingredient_usage_report_api_v1_reports_ingredient_usage_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Start date for the report period (YYYY-MM-DD)",
              "title": "Start Date"
            },
            "description": "Start date for the report period (YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "End date for the report period (YYYY-MM-DD)",
              "title": "End Date"
            },
            "description": "End date for the report period (YYYY-MM-DD)"
          },
          {
            "name": "output_format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Output format for the report",
              "enum": [
                "json",
                "csv",
                "pdf"
              ],
              "default": "json",
              "title": "Output Format"
            },
            "description": "Output format for the report"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Ingredient Usage Report Api V1 Reports Ingredient Usage Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/low-stock": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Low Stock Report",
        "description": "Generate a Low Stock report for ingredients below their threshold.\nOutput can be JSON, CSV, or PDF (PDF is placeholder).",
        "operationId": "get_low_stock_report_api_v1_reports_low_stock_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "output_format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Output format for the report",
              "enum": [
                "json",
                "csv",
                "pdf"
              ],
              "default": "json",
              "title": "Output Format"
            },
            "description": "Output format for the report"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Low Stock Report Api V1 Reports Low Stock Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login/access-token"
          }
        }
      }
    }
  }
}